# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a todo-list server.",
        "title": "Gin Swagger todo-list API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/tasks": {
            "get": {
                "description": "Gets all tasks in database that pass applied filter",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "getTasksByFilter",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "complete",
                        "name": "complete",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Adds new task at the end of database",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "postTasks",
                "parameters": [
                    {
                        "description": "Task to add",
                        "name": "Task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.TaskPayload"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Deletes all tasks",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "deleteTasks",
                "responses": {}
            }
        },
        "/tasks/id/{id}": {
            "delete": {
                "description": "Deletes tast at specified ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "deleteTaskByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The specified ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "Toggles complete at specified ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "patchCompleteByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The specified ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.TaskPayload": {
            "type": "object",
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}